dataCollectingWebApp/backend/requirements.txt
django-rest-framework
django-cors-headers
django-instagram
google-api-python-client
google-images-search